Fix 9 unit test failures across 5 modules
1. Parquet serialization roundtrip (4 checkpoint + 1 freshness test):
   - CheckpointStore/FreshnessStore.load_all() now uses read_parquet_or_ndjson()
     to find NDJSON fallback files written by save_dataframe_parquet()
   - Fixed dtype check: `== "object"` → `not is_datetime64_any_dtype()` to
     handle pandas 2.2's StringDtype (was silently skipping datetime conversion)
   - save_dataframe_parquet() now removes stale .parquet when falling back to
     NDJSON, preventing read_parquet_or_ndjson from reading outdated data
   - get_stale_records/get_record: NaN→None cleanup before Pydantic validation

2. CET aggregator (awards_with_cet==3 instead of 2):
   - _extract_cet_rows_from_award() now guards against NaN primary_cet
     (pandas stores None as NaN in object columns, and NaN is truthy in Python)

3. Transition analytics (denominator==1 instead of 3):
   - _company_id_series() uses .fillna("") before .astype(str) to prevent
     NaN propagation through string concatenation ("uei:" + NaN = NaN)

4. Task parser (empty file_hints):
   - _FILE_HINT_PATTERN regex: `src/` prefix → any path ending in .py

5. r_stateio_functions (int64 dtype error):
   - Cast industry_output to float before zero-replacement assignment

+++ sbir_etl/transformers/r_stateio_functions.py
@@ -470,8 +470,8 @@ def calculate_technical_coefficients(
         )
 
     # Avoid division by zero
-    # Replace zero outputs with small epsilon
-    output_safe = industry_output.copy()
+    # Replace zero outputs with small epsilon (cast to float to avoid int64 dtype issues)
+    output_safe = industry_output.astype(float).copy()
     output_safe[output_safe == 0] = 1e-10
 
     # Calculate technical coefficients: A[i,j] = Use[i,j] / Output[j]
